experiment_runtime/pyproject.toml
[project]
name = "fhe-experiment-runtime"
version = "0.1.0"
description = "LangGraph scaffold for the multi-agent FHE privacy experiment."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=1.0.0",
    "langgraph>=0.6.0",
    "openai>=2.0.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]

[project.scripts]
fhe-experiment-run = "qwen_langgraph_demo.main:main"
fhe-endpoint-smoke = "qwen_langgraph_demo.test_endpoint:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/qwen_langgraph_demo",
    "src/experiment",
]
